Fix bug where overriding symlink to a dir created cyclic symlinks
Creating a symlink to a directory works as expected when there is no pre-existing directory with the same name as `fq_dest`. If the `fq_dest` is a directory and already exists, then `ln` will follow the existing link and create a link to `fq_src`, resulting in a cyclic symlink.

E.g. Creating a symlink from a non-existent `dest_dir` to `src_dir`:
```
.
├── dest_dir -> src_dir
└── src_dir
```

Overriding a symlink from a pre-existing`dest_dir` to `src_dir`:
```
.
├── dest_dir -> src_dir
└── src_dir
    └── src_dir -> src_dir
```